move ranges
