+++ ydb/library/range_treap/range_treap.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ydb/library/accessor/accessor.h>
+#include "range.h"
 
 #include <util/generic/hash.h>
 #include <util/generic/intrlist.h>
@@ -8,138 +8,6 @@
 
 namespace NKikimr {
 namespace NRangeTreap {
-
-template <class TKey>
-struct TOwnedRange {
-    TKey LeftKey;
-    TKey RightKey;
-    bool LeftInclusive;
-    bool RightInclusive;
-
-    TOwnedRange(TKey leftKey, bool leftInclusive, TKey rightKey, bool rightInclusive) noexcept
-        : LeftKey(std::move(leftKey))
-        , RightKey(std::move(rightKey))
-        , LeftInclusive(leftInclusive)
-        , RightInclusive(rightInclusive)
-    {
-    }
-};
-
-template <class TKey, class TKeyView>
-struct TRange {
-    TKeyView LeftKey;
-    TKeyView RightKey;
-    bool LeftInclusive;
-    bool RightInclusive;
-
-    TRange(TKeyView leftKey, bool leftInclusive, TKeyView rightKey, bool rightInclusive) noexcept
-        : LeftKey(leftKey)
-        , RightKey(rightKey)
-        , LeftInclusive(leftInclusive)
-        , RightInclusive(rightInclusive)
-    {
-    }
-
-    TOwnedRange<TKey> ToOwnedRange() const {
-        TKey leftKey(LeftKey);
-        TKey rightKey;
-        if (LeftKey.data() != RightKey.data() || LeftKey.size() != RightKey.size()) {
-            rightKey = TOwnedCellVec(RightKey);
-        } else {
-            rightKey = leftKey;
-        }
-        return TOwnedRange(std::move(leftKey), LeftInclusive, std::move(rightKey), RightInclusive);
-    }
-};
-
-enum EBorderMode: ui8 {
-    LeftExclusive = 0,
-    LeftInclusive = 1,
-    RightExclusive = 2,
-    RightInclusive = 3,
-};
-
-class TBorderModeTraits {
-public:
-    static bool IsInclusive(EBorderMode mode) noexcept {
-        return mode & FlagInclusive;
-    }
-
-    static bool IsRight(EBorderMode mode) noexcept {
-        return mode & FlagRight;
-    }
-
-    static std::strong_ordering CompareEqualPoint(const EBorderMode lhs, const EBorderMode rhs) {
-        if (lhs == rhs) {
-            return std::strong_ordering::equal;
-        }
-        if (!IsInclusive(lhs)) {
-            if (IsRight(lhs)) {
-                return std::strong_ordering::greater;
-            } else {
-                return std::strong_ordering::less;
-            }
-        }
-        if (!IsInclusive(rhs)) {
-            if (IsRight(rhs)) {
-                return std::strong_ordering::less;
-            } else {
-                return std::strong_ordering::greater;
-            }
-        }
-        if (IsRight(rhs)) {
-            return std::strong_ordering::less;
-        } else {
-            return std::strong_ordering::greater;
-        }
-    }
-
-private:
-    static const ui8 FlagInclusive = (1 << 0);
-    static const ui8 FlagRight = (1 << 1);
-};
-
-template <class TKeyView>
-class TBorder {
-private:
-    TKeyView Key;
-    YDB_READONLY_DEF(EBorderMode, Mode);
-
-public:
-    static TBorder MakeLeft(TKeyView key, bool inclusive) noexcept {
-        return { key, inclusive ? EBorderMode::LeftInclusive : EBorderMode::LeftExclusive };
-    }
-
-    static TBorder MakeRight(TKeyView key, bool inclusive) noexcept {
-        return { key, inclusive ? EBorderMode::RightInclusive : EBorderMode::RightExclusive };
-    }
-
-    const TKeyView& GetKey() const {
-        return Key;
-    }
-
-    TBorder(const TKeyView& key, const EBorderMode mode)
-        : Key(key)
-        , Mode(mode)
-    {
-    }
-};
-
-template <class TKeyView>
-class TDefaultBorderComparator {
-    using TBorder = TBorder<TKeyView>;
-
-public:
-    static std::strong_ordering Compare(const TBorder& lhs, const TBorder& rhs) {
-        if (lhs.GetKey() < rhs.GetKey()) {
-            return std::strong_ordering::less;
-        }
-        if (lhs.GetKey() > rhs.GetKey()) {
-            return std::strong_ordering::greater;
-        }
-        return TBorderModeTraits::CompareEqualPoint(lhs.GetMode(), rhs.GetMode());
-    }
-};
 
 template <class TKey, class TValue>
 struct TDefaultValueTraits {
@@ -439,8 +307,8 @@ private:
 
 private:
     /**
-          * Removes node t from the tree
-          */
+      * Removes node t from the tree
+      */
     void DoRemove(TNode* t) {
         Y_DEBUG_ABORT_UNLESS(t, "Trying to remove a nullptr node");
         TNode* p = t->Parent;
@@ -458,8 +326,8 @@ private:
     }
 
     /**
-          * Removes the node linked by tptr from the tree
-          */
+      * Removes the node linked by tptr from the tree
+      */
     void DoRemove(THolder<TNode>* tptr) {
         THolder<TNode> d = std::move(*tptr);
         Y_DEBUG_ABORT_UNLESS(d, "Cannot remove a null node");
@@ -498,10 +366,10 @@ private:
 
 private:
     /**
-          * Recomputes MaxRightKey for subtree root t
-          *
-          * Returns the node which was the source of the new MaxRightKey
-          */
+      * Recomputes MaxRightKey for subtree root t
+      *
+      * Returns the node which was the source of the new MaxRightKey
+      */
     TNode* RecomputeMaxRight(TNode* t) {
         TNode* source = t;
         t->MaxRightKey = t->RightKey;
@@ -523,8 +391,8 @@ private:
     }
 
     /**
-          * Recomputes MaxRightKey for subtree root t and its parents
-          */
+      * Recomputes MaxRightKey for subtree root t and its parents
+      */
     void RecomputeMaxRights(TNode* t) {
         while (t) {
             RecomputeMaxRight(t);
@@ -533,10 +401,10 @@ private:
     }
 
     /**
-          * Extends RightKey of node t if necessary
-          *
-          * Returns true when t->MaxRightKey is modified
-          */
+      * Extends RightKey of node t if necessary
+      *
+      * Returns true when t->MaxRightKey is modified
+      */
     bool ExtendRightKey(TNode* t, const TBorder& rightKey, TKey rightOwnedKey) {
         if (CompareBorders(t->RightBorder(), rightKey) < 0) {
             t->RightKey = rightOwnedKey;
@@ -560,10 +428,10 @@ private:
     }
 
     /**
-          * Extends MaxRightKey of node t
-          *
-          * Returns true when t->MaxRightKey is modified
-          */
+      * Extends MaxRightKey of node t
+      *
+      * Returns true when t->MaxRightKey is modified
+      */
     bool ExtendMaxRightKey(TNode* t, const TKey& rightKey, EBorderMode rightMode) {
         int cmp = CompareBorders(t->MaxRightBorder(), TBorder(rightKey, rightMode));
         if (cmp < 0) {
@@ -577,8 +445,8 @@ private:
     }
 
     /**
-          * Extends MaxRightKey of node t and all its parents
-          */
+      * Extends MaxRightKey of node t and all its parents
+      */
     void ExtendMaxRightKeys(TNode* t, const TKey& rightKey, EBorderMode rightMode) {
         while (t && ExtendMaxRightKey(t, rightKey, rightMode)) {
             t = t->Parent;
@@ -636,8 +504,8 @@ private:
 
 public:
     /**
-          * Returns height of tree, for tests only, O(n) complexity
-          */
+      * Returns height of tree, for tests only, O(n) complexity
+      */
     size_t Height() const {
         return Root ? DoCalcHeight(Root.Get()) : 0;
     }
@@ -669,8 +537,8 @@ public:
 
 private:
     /**
-          * Validates all invariants for subtree t
-          */
+      * Validates all invariants for subtree t
+      */
     std::tuple<TNode*, TNode*> DoValidate(TNode* t) const {
         int cmp;
         TNode* leftMost = t;
